refactor: 코어 로직 관련 네이밍 컨벤션 변경 (core/repositories -> core/infrastructure)
- 외부 저장소에서 데이터를 받아온 후, 서비스 로직을 거쳐 프로그램에서 활용되느 자료구조(entity)로 변경하는 core 로직 관련하여 네이밍 컨벤션 변경
- auth token 등도 외부 저장소에서 받아오는 데이터이지만, repository로 분류하기엔 애매한 부분이 있어서 보다 포괄적인 infrastructure로 변경
