+++ .gitignore
@@ -40,6 +40,7 @@ yarn-error.log*
 *.tsbuildinfo
 next-env.d.ts
 
-## md
+## docs
 *.md
+*.pdf
