+++ src/core/infrastructure/_security.yfinance.infra.ts
